+++ src/components/title1/Title1TreeMap.tsx
@@ -211,7 +211,7 @@ export default function Title1TreeMap({ program, TreeMapData, year, stateCodes, 
                         >
                             <i>
                                 The payments,base acres and payment recipients are calculated as the total of the data
-                                from 2014-2021. 2022 payments for Title I have not yet been paid.
+                                from 2014-2023. 2024 payments for Title I have not yet been paid.
                             </i>
                         </Typography>
                         <Typography